@@ -1,49 +1,85 @@
-<template>
-  <div className="App">
-    <p>You clicked me {{ count }} times.</p>
-    <button @click="onClick()">Click Me!</button>
-  </div>
-</template>
-
-<script>
-export default {
-  data() {
-    return {
-      count: 0,
-    };
-  },
-  methods: {
-    onClick() {
-      fetch(process.env.VUE_APP_API_URL, {
-        method: "POST",
-      })
-        .then((response) => response.text())
-        .then((data) => {
-          this.count = data;
-        });
-    },
-  },
-};
+<script setup lang="ts">
+import { RouterLink, RouterView } from 'vue-router'
+import HelloWorld from './components/HelloWorld.vue'
 </script>
 
-<style>
-body,
-html {
-  height: 100vh;
-  display: grid;
-  place-items: center;
+<template>
+  <header>
+    <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
+
+    <div class="wrapper">
+      <HelloWorld msg="You did it!" />
+
+      <nav>
+        <RouterLink to="/">Home</RouterLink>
+        <RouterLink to="/about">About</RouterLink>
+      </nav>
+    </div>
+  </header>
+
+  <RouterView />
+</template>
+
+<style scoped>
+header {
+  line-height: 1.5;
+  max-height: 100vh;
 }
-#root {
-  margin: auto;
+
+.logo {
+  display: block;
+  margin: 0 auto 2rem;
 }
-.App {
+
+nav {
+  width: 100%;
+  font-size: 12px;
   text-align: center;
+  margin-top: 2rem;
 }
-p {
-  margin-top: 0;
-  font-size: 20px;
+
+nav a.router-link-exact-active {
+  color: var(--color-text);
 }
-button {
-  font-size: 48px;
+
+nav a.router-link-exact-active:hover {
+  background-color: transparent;
+}
+
+nav a {
+  display: inline-block;
+  padding: 0 1rem;
+  border-left: 1px solid var(--color-border);
+}
+
+nav a:first-of-type {
+  border: 0;
+}
+
+@media (min-width: 1024px) {
+  header {
+    display: flex;
+    place-items: center;
+    padding-right: calc(var(--section-gap) / 2);
+  }
+
+  .logo {
+    margin: 0 2rem 0 0;
+  }
+
+  header .wrapper {
+    display: flex;
+    place-items: flex-start;
+    flex-wrap: wrap;
+  }
+
+  nav {
+    text-align: left;
+    margin-left: -1rem;
+    font-size: 1rem;
+
+    padding: 1rem 0;
+    margin-top: 1rem;
+  }
 }
 </style>
